finalization transfert to fastapi
