+++ activetigger/project.py
@@ -1,8 +1,7 @@
-"""
-Classes available : Projet, Features, Schemes 
-"""
 import os
+from datetime import datetime
 from pathlib import Path
+import sqlite3
 import yaml # type: ignore
 import pandas as pd # type: ignore
 import re
@@ -10,158 +9,263 @@ import pyarrow.parquet as pq # type: ignore
 import json
 import functions
 from models import SimpleModel, BertModel
+from datamodels import ParamsModel, SchemesModel, SchemeModel, SimpleModelModel, BertModelModel
 from pandas import DataFrame, Series
 from pydantic import BaseModel
+from fastapi import UploadFile
+from fastapi.encoders import jsonable_encoder
 
 import logging
 logging.basicConfig(filename='log.log', 
                     encoding='utf-8', 
                     level=logging.INFO,
                     format='%(asctime)s - %(levelname)s - %(message)s')
-
-
-# optimiser le calcul des embeddings avec GPU
-# lancer un processus indépendant pour calculer les embeddings
-# use Pydantic & BaseModel to type data in API and automatically
-# generate the JSON content https://realpython.com/api-integration-in-python/ 
-
-# plus précisémment définir le format des objets qui circulent entre le back et le front
-# - element
-# - options
-# - liste d'éléments
-
-# TODO : gérer la sauvegarde de tous les éléments
-# - schemes gérés avec les paramètres (mais ça doit changer)
-# - features gestion à part
-# - data + coding
-
-# TODO : gérer les éléments tagged dans next etc.
-
-class Project():
+class Session():
     """
-    Project (database/params)
+    Global session parameters
+    TODO : load from yaml file
     """
-    def __init__(self, project_name : str, **kwargs) -> None:
-        """
-        Initialize a project 
-        (load or create)
-        """
+    max_file_size: int = 50 * 1024 * 1024
+    path: Path = Path("../projects")
+    db: Path = Path('../projects/activetigger.db')
+    features_file:str = "features.parquet"
+    labels_file:str = "labels.parquet"
+    data_file:str = "data.parquet"
 
-        self.name: str = project_name
-        self.schemes: Schemes = Schemes(self.name)
-        self.features = Features(project_name=self.name)
-        self.params: dict = {}
-        self.content: DataFrame = DataFrame()
-        self.simplemodel:SimpleModel = SimpleModel()
-        self.bertmodel:BertModel = BertModel(path = f"{self.name}")
 
-        # load or create project    
-        if self.exists(project_name):
-             self.load_params()
-             self.load_data()
-             self.features.load()
+class Server(Session):
+    """
+    Projects manager managing fastapi requests
+    """
+    def __init__(self) -> None:
+        """
+        Start the server
+        """
+        logging.info("Starting server")
+        logging.warning('Still under development')
+
+        self.projects: dict = {}
+        self.time_start = datetime.now()
+
+        if not self.db.exists():
+            logging.info("Creating database")
+            self.create_db()
+
+    def create_db(self):
+        """
+        Initialize the database
+        """
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+
+        # Projects table
+        create_table_sql = '''
+            CREATE TABLE IF NOT EXISTS projects (
+                project_name TEXT PRIMARY KEY,
+                time_created TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+                parameters TEXT,
+                time_modified TIMESTAMP,
+                user TEXT
+            )
+        '''
+        cursor.execute(create_table_sql)
+
+        # Schemes table
+        create_table_sql = '''
+            CREATE TABLE IF NOT EXISTS schemes (
+                id INTEGER PRIMARY KEY AUTOINCREMENT,
+                time_created TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+                time_modified TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+                user TEXT,
+                project TEXT,
+                name TEXT,
+                params TEXT
+            )
+        '''
+        cursor.execute(create_table_sql)
+
+        # History table
+        # Users table
+
+        conn.commit()
+        conn.close()
+        return conn
+
+    def db_get_project(self, project_name:str) -> ParamsModel|None:
+        """
+        Get project from database
+        """
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "SELECT * FROM projects WHERE project_name = ?"
+        cursor.execute(query, (project_name,))
+        existing_project = cursor.fetchone()
+        conn.commit()
+        conn.close()
+
+        if existing_project:
+            p = ParamsModel(**json.loads(existing_project[2]))
+            return p
         else:
-             self.create(project_name, **kwargs)
+            return None
+
+    def exists(self, project_name):
+        existing = self.get_existing_projects()
+        v = (project_name in existing)
+        return v
+
+    def get_existing_projects(self):
+        """
+        Projects already existing
+        """
+        return [i for i in os.listdir(self.path) if os.path.isdir(self.path / i)]
+
+    def start_project(self,project_name):
+        """
+        Load project
+        """
+        # The project has to exist
+        logging.info(f"Load project {project_name}")
+        self.projects[project_name] = Project(project_name)
+
+    def update_project_parameters(self, project):
+        """
+        Update project parameters in the DB
+        """
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "SELECT * FROM projects WHERE project_name = ?"
+        cursor.execute(query, (project.project_name,))
+        existing_project = cursor.fetchone()
+
+        if existing_project:
+            # Update the existing project
+            update_query = "UPDATE projects SET parameters = ?, time_modified = CURRENT_TIMESTAMP WHERE project_name = ?"
+            cursor.execute(update_query, (json.dumps(jsonable_encoder(project)), project.project_name))
+        else:
+            # Insert a new project
+            insert_query = "INSERT INTO projects (project_name, parameters, time_modified) VALUES (?, ?, CURRENT_TIMESTAMP)"
+            cursor.execute(insert_query, (project.project_name, json.dumps(jsonable_encoder(project))))
+        conn.commit()
+        conn.close()
+        return True
+
+    def create_project(self, 
+                       params:ParamsModel, 
+                       file:UploadFile) -> ParamsModel:
+        """
+        Set up a new project
+        - load data and save
+        - initialize parameters
+        - initialize files
+
+        TODO : move to parquet files
+        """
+        # create directory
+        params.dir = self.path / params.project_name
+        os.makedirs(params.dir)
+
+        # write data
+        with open(params.dir / "data_raw.csv","wb") as f:
+            f.write(file.file.read())
+
+        # parameters of projects
+        if params.col_text is None:
+            params.col_text = "text"
+        if not params.col_tags is None: # if existing tag column
+            print("Not implemented")
+        if params.n_rows is None:
+            params.n_rows = 2000 # limit of lines if not given
+        # TODO : Question of embeddings
+
+        # save parameters 
+        self.update_project_parameters(params)
+
+        # load only the number of rows for the project
+        content = pd.read_csv(params.dir / "data_raw.csv", 
+                                index_col=0, 
+                                nrows=params.n_rows)
+        
+        content.index = [str(i) for i in content.index]
+        print(content.index)
     
-    def exists(self, project_name: str) -> bool:
+        # create the empty annotated file / features file
+        content.to_parquet(params.dir / self.data_file, index=True)
+        content[[params.col_text]].to_parquet(params.dir / self.labels_file, index=True)
+        content[[]].to_parquet(params.dir / self.features_file, index=True)
+
         """
-        Test if a project exists
-        """
-        return Path(f"{project_name}/{project_name}.yaml").exists()
-
-    def create(self, project_name:str, **kwargs):
-        """
-        Create new project
-        """
-
-        ## TO DO : tests que tous les paramètres ont bien été renseignés
-        # Le premier schéma enregistré est appelé default
-        
-        if not Path(project_name).exists():
-            os.makedirs(project_name)
-        else:
-            print("Erreur, le dossier existe") # gérer la gestion des erreurs
-
-        # Load data
-        self.content = pd.read_csv(kwargs["file"],
-                                   index_col=0,
-                                   low_memory=False,
-                                   nrows=kwargs["n_rows"])
-
-        # Manage parameters
-        if not "col_text" in kwargs:
-            kwargs["col_text"] = "text"
-        if "col_tags" in kwargs: # if existing tag column
-            kwargs["cat"] = list(self.content[kwargs["col_tags"]].dropna().unique())
-        if not "sbert" in kwargs:
-            kwargs["sbert"] = False
-        if not "fasttext" in kwargs:
-            kwargs["fasttext"] = False
-
-        # TODO : test que les colonnes utilisées existent bien
-            
-        self.params = {
-                    "project_name":project_name,
-                    "origin_file":kwargs["file"],
-                    "n_rows":kwargs["n_rows"],
-                    "col_text":kwargs["col_text"],
-                    "col_tags":"labels_default",
-                    "scheme":"default",
-                    "cat":{"default":kwargs["cat"]},
-                    "embeddings":{
-                        "sbert":kwargs["sbert"],
-                        "fasttext":kwargs["fasttext"],
-                    }
-                  }
-        
-        self.schemes.add("default",kwargs["cat"])
-        self.schemes.select("default")
-
+        TODO : deal if already tagged column
         if "col_tags" in kwargs: # if existing tag column
             self.content[self.schemes.col] = self.content[kwargs["col_tags"]]
         else:
             self.content[self.schemes.col] = None
+        """
 
-        # Add / compute embeddings as features
-        if self.params["embeddings"]["sbert"]:
+        return params
+    
+
+
+class Project(Session):
+    """
+    Project object
+    """
+
+    def __init__(self, 
+                 project_name: str) -> None:
+        """
+        Load a project
+        """
+        self.name: str = project_name
+        self.params: ParamsModel = self.load_params(project_name)
+        self.content: DataFrame = pd.read_parquet(self.params.dir / self.data_file)
+        self.schemes: Schemes = Schemes(project_name, 
+                                        self.params.dir / self.labels_file)
+        self.features: Features = Features(project_name,
+                                           self.params.dir / self.features_file)
+        self.bertmodel: BertModel = BertModel(self.params.dir)
+        self.simplemodel: SimpleModel = SimpleModel()
+
+        # Compute features if requested
+        if ("sbert" in self.params.embeddings) & (not "sbert" in self.features.map):
             self.features.add("sbert",
                               self.compute_embeddings(emb="sbert"))
-        if self.params["embeddings"]["fasttext"]:
+        if ("fasttext" in self.params.embeddings) & (not "fasttext" in self.features.map):
             self.features.add("fasttext",
                               self.compute_embeddings(emb="fasttext"))
 
-        self.save_params()
-        self.save_data()
-    
+    def load_params(self, project_name:str) -> ParamsModel:
+        """
+        Load params from database
+        """
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "SELECT * FROM projects WHERE project_name = ?"
+        cursor.execute(query, (project_name,))
+        existing_project = cursor.fetchone()
+        conn.commit()
+        conn.close()
+
+        if existing_project:
+            return ParamsModel(**json.loads(existing_project[2]))
+        else:
+            raise NameError(f"{project_name} does not exist.")
+
+
     def compute_embeddings(self,
-                           emb:None|str = None):
+                           emb:str) -> DataFrame:
         """
-        Compute embeddings (and save it)
+        Compute embeddings
+        TODO : async ?
         """
+        print("start compute embeddings ",emb, self.content.shape)
         if emb == "fasttext":
-            file = f"{self.name}/fastext"
-            if Path(file).exists():
-                print("Fasttext embeddings already exist")
-                emb_fasttext = pd.read_csv(file,index_col=0)
-            else:
-                print("Starting to compute fasttext embeddings")
-                emb_fasttext = functions.to_fasttext(self.content[self.params["col_text"]])
-                emb_fasttext.to_csv(file)
-            self.params["embeddings"]["fasttext"] = True
-            self.save_params()
+            emb_fasttext = functions.to_fasttext(self.content[self.params.col_text])
             return emb_fasttext
         if emb == "sbert":
-            file = f"{self.name}/sbert"
-            if Path(file).exists():
-                print("Sbert embeddings already exist")
-                emb_sbert = pd.read_csv(file,index_col=0)
-            else:
-                print("Starting to compute sbert embeddings")
-                emb_sbert = functions.to_sbert(self.content[self.params["col_text"]])
-                emb_sbert.to_csv(file)
-            self.params["embeddings"]["sbert"] = True
-            self.save_params()
+            emb_sbert = functions.to_sbert(self.content[self.params.col_text])
             return emb_sbert
+        raise NameError(f"{emb} does not exist.")
     
     def fit_simplemodel(self,
                         model:str,
@@ -177,7 +281,7 @@ class Project():
         df_features = self.features.get(features)
         col_label = self.schemes.col
         col_predictors = df_features.columns
-        data = pd.concat([self.content[col_label],
+        data = pd.concat([self.schemes.content[col_label],
                           df_features],
                           axis=1)
         
@@ -189,31 +293,15 @@ class Project():
                         )
         return s
     
-    def update_simplemodel(self, req:dict) -> dict:
-        if req["features"] is None or len(req["features"])==0:
+    def update_simplemodel(self, simplemodel: SimpleModelModel) -> dict:
+        if simplemodel.features is None or len(simplemodel.features)==0:
             return {"error":"no features"}
         self.simplemodel = self.fit_simplemodel(
-                                model=req["current"],
-                                features=req["features"],
-                                model_params=req["parameters"]
+                                model=simplemodel.current,
+                                features=simplemodel.features,
+                                model_params=simplemodel.params
                                 )
         return {"success":"new simplemodel"}
-
-    def load_params(self):
-        """
-        Load YAML configuration file
-        """
-        with open(f"{self.name}/{self.name}.yaml","r") as f:
-            self.params =  yaml.safe_load(f)
-
-        # load also schemes (TO CHANGE IN THE FUTURE)
-        self.schemes = Schemes(self.name)
-        self.schemes.load({
-            "project_name":self.name,
-            "name":self.params["scheme"],
-            "labels":self.params["cat"][self.params["scheme"]],
-            "available":self.params["cat"]
-            })
 
     def save_params(self,params: None|dict=None) -> None:
         """
@@ -231,46 +319,29 @@ class Project():
         self.schemes.load(json)
         self.params["cat"] = self.schemes.available
         self.save_params()
-
-    def load_data(self):
-        """
-        Load data
-        """
-        self.content = pd.read_csv(f"{self.name}/{self.name}.csv",
-                                   index_col=0,
-                                   low_memory=False
-                                   )
-    def save_data(self):
-        """
-        Save data
-        """
-        self.content.to_csv(f"{self.name}/{self.name}.csv")
-
-    def save(self):
-        """
-        Save all
-        """
-        # data
-        self.content.to_csv(f"{self.name}/{self.name}.csv")
-        # params
-        self.save_params()
-        # features
-        self.features.save()
-        # simplemodels not saved
     
     def delete_label(self,element_id):
         """
         Delete a recorded tag
         """
-        self.content.loc[element_id,self.schemes.col] = None
+        self.schemes.content.loc[element_id,self.schemes.col] = None
         return True
 
-    def add_label(self,element_id,label):
+    def add_label(self,element_id, tag, scheme = None):
         """
         Record a tag
         """
-        self.content.loc[element_id,self.schemes.col] = label
-        return True
+        if scheme is None: #current sheme
+            scheme = self.schemes.col
+
+        if not scheme in self.schemes.available():
+            return {"error":"scheme unavailable"}
+
+        self.schemes.content.loc[element_id,scheme] = tag
+
+        # TODO : ajouter dans l'historique qui a modifié le tag
+
+        return {"success":"tag added"}
 
     def get_next(self,
                  scheme:str,
@@ -290,12 +361,12 @@ class Project():
             self.schemes.select(scheme)
 
         # Pour le moment uniquement les cases non nulles
-        f = self.content[self.schemes.col].isnull()
+        f = self.schemes.content[self.schemes.col].isnull()
 
         if mode == "deterministic": # next row
-            element_id = self.content[f].index[0]
+            element_id = self.schemes.content[f].index[0]
         if mode == "random": # random row
-            element_id = self.content[f].sample(random_state=42).index[0]
+            element_id = self.schemes.content[f].sample(random_state=42).index[0]
         if mode == "maxprob": # higher prob row
             if self.simplemodel.name is None: # if no model, build default
                 print("Build default simple model")
@@ -311,11 +382,7 @@ class Project():
         # TODO : put a lock on the element when sent ?
 
         # Pour le moment uniquement l'id et le texte (dans le futur ajouter tous les éléments)
-        return  {
-                 "element_id":element_id,
-                 "content":self.get_element(element_id),
-#                 "options":self.get_state()
-                }
+        return  self.get_element(element_id)
     
     def get_element(self,element_id):
         """
@@ -325,7 +392,6 @@ class Project():
         return {"id":element_id,
                 "text":self.content.loc[element_id,"text"]
                 }
-
 
     def get_params(self):
         """
@@ -353,164 +419,240 @@ class Project():
                         },
                     "scheme":{
                                 "current":self.schemes.name,
-                                "available":self.schemes.available
+                                "available":self.schemes.available()
                                 },
                     "features":{
-                            "available_features":self.features.available
+                            "available_features":self.features.map.keys()
                           }
                    }
         
-
-
-        
-        return  {
-                 "type":"state",
-                 "content":options
-                }
+        return  options
     
     def add_regex(self, name: str, value: str):
         """
         Add regex to features
         """
-        if not name in self.features.available:
-            pattern = re.compile(value)
-            f = self.content[self.params["col_text"]].apply(lambda x: bool(pattern.search(x)))
-            self.features.add(name,f)
-            print(self.features.available)
-            return {"success":"added"}
-        else:
-            return {"error":"exists already"}
-
-class Features():
+        pattern = re.compile(value)
+        f = self.content[self.params.col_text].apply(lambda x: bool(pattern.search(x)))
+        self.features.add(name,f)
+    
+class Features(Session):
     """
-    Project features
-    No duplicate of data
-    TODO : test for the length of the data/same index
-    TODO : load available features
+    Manage project features
+    Comment : as a file
     """
 
-    def __init__(self, project_name:str) -> None:
+    def __init__(self, 
+                 project_name:str, 
+                 data_path:Path) -> None:
         self.project_name = project_name
-        self.available:list = []
-        self.map:dict = {}
-        self.content = None
+        self.path = data_path
+
+        # load data from file
+        content, map = self.load()
+        self.content: DataFrame = content
+        self.map:dict = map
 
     def __repr__(self) -> str:
-        return f"Available features : {self.available}"
+        return f"Available features : {self.map}"
     
-    def save(self):
+    def load(self):
         """
-        Save current state of embeddings
-        Temporary : CSV
+        Load file and agregate columns
         """
-        metadata = {
-            "project_name":self.project_name,
-            "available":self.available,
-            "map":self.map
-        }
-
-        table = self.content.copy()
-        table["params"] = None
-        table["params"].iloc[0] = json.dumps(metadata)
-        table.to_csv(f"{self.project_name}/features.csv")
-
-    def load(self) -> bool:
-        """
-        Load existing features
-        Temporary : CSV
-        """
-        if Path(f"{self.project_name}/features.csv").exists():
-            table = pd.read_csv(f"{self.project_name}/features.csv", 
-                                index_col=0,
-                                low_memory=False)
-            metadata = json.loads(table["params"].iloc[0])
-            self.content = table.drop(columns="params").copy()
-            del table
-            self.map = metadata["map"]
-            self.available = metadata["available"]
-            self.project_name = metadata["project_name"]
-            return True
-        else:
-            return False
-
+        def find_strings_with_pattern(strings, pattern):
+            matching_strings = [s for s in strings if re.match(pattern, s)]
+            return matching_strings
+        data = pd.read_parquet(self.path)
+        var = set([i.split("_")[0] for i in data.columns])
+        dic = {i:find_strings_with_pattern(data.columns,i) for i in var}
+        return data, dic
+    
     def add(self, 
             name:str, 
-            content:DataFrame|Series) -> None:
+            content:DataFrame|Series) -> dict:
         """
-        Add feature(s)
+        Add feature(s) and save
         """
+
+        # test length
+        if len(content) != len(self.content):
+            raise ValueError("Features don't have the right shape") 
+        
+        if name in self.map:
+            return {"error":"feature name already exists"}
+
+        # change type
         if type(content)==Series:
             content = pd.DataFrame(content)
+
+        # add to the table & dictionnary
         content.columns = [f"{name}_{i}" for i in content.columns]
         self.map[name] = list(content.columns)
-        # create the dataset
-        if self.content is None:
-            self.content = content
-        else:
-            self.content = pd.concat([self.content,content],
+        self.content = pd.concat([self.content,content],
                                      axis=1)
-        self.available.append(name)
+        # save
+        self.content.to_parquet(self.path)
+        return {"success":"feature added"}
 
     def delete(self, name:str):
         """
         Delete feature
         """
-        if name in self.available:
-            col = self.get([name])
-            self.available.remove(name)
-            self.content.drop(columns=col)
-            return {"success":"feature deleted"}
-        else:
+        if not name in self.map:
             return {"error":"feature doesn't exist"}
+
+        col = self.get([name])
+        del self.map[name]
+        self.content.drop(columns=col)
+        self.content.to_csv(self.path,index=True)
+        return {"success":"feature deleted"}
             
-    def get(self,features:list|str = "all"):
+    def get(self, features:list|str = "all"):
         """
         Get specific features
         """
-        cols = []
         if features == "all":
-            features = self.available
+            features = list(self.map.keys())
+        if type(features) is str:
+            features = [features]
 
+        cols = []
         for i in features:
-            if i in self.available:
+            if i in self.map:
                 cols += self.map[i]
-            else:
-                print(f"Feature {i} doesn't exist")
 
         return self.content[cols]
-    
 
-class Schemes():
+class Schemes(Session):
     """
-    Project Schemes
+    Manage project schemes & tags
     """
-    def __init__(self,project_name):
+    def __init__(self,
+                 project_name: str,
+                 path:Path):
+        """
+        Init empty
+        """
         self.project_name = project_name
+        self.path = path
         self.name = None
         self.labels = None
         self.col = None
-        self.available = {}
+
+        # Load data
+        self.content = pd.read_parquet(self.path)
+
+        # Initialize the current scheme
+        available = self.available()
+        if len(available) == 0:
+            self.add("default",[])
+            self.select("default")
+        else:
+            self.select(list(available.keys())[0])
 
     def __repr__(self) -> str:
-        return f"Coding schemes available {self.available}"
+        return f"Coding schemes available {self.available()}"
 
     def col_name(self):
-        return "labels_" + self.name
+        return self.name
 
     def select(self, name):
-        if name in self.available:
+        available = self.available()
+        if name in available:
             self.name = name
-            self.labels = self.available[name]
+            self.labels = available[name]
             self.col = self.col_name()
         else:
             raise IndexError
 
-    def add(self, name: str, modalities: list):
-        if not name in self.available:
-            self.available[name] = modalities
-            return True
+    def add(self, scheme:SchemeModel):
+        """
+        Add new scheme
+        - in database
+        - in file
+        """
+        if self.exists(scheme.name):
+            return {"error":"scheme name already exists"}
+        
+        # add it if in database
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "INSERT INTO schemes (project, name, params) VALUES (?, ?, ?)"
+        cursor.execute(query, (self.project_name, scheme.name, json.dumps(scheme.tags)))
+        conn.commit()
+        conn.close()
+
+        # add it in the file
+        self.content[scheme.name] = None
+        self.content.to_parquet(self.path)
+
+        return {"success":"scheme created"}
+
+    def update(self, scheme:SchemeModel):
+        """
+        Update existing schemes from database
+        """
+        if not self.exists(scheme.name):
+            return {"error":"scheme doesn't exist in db"}
+        
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "UPDATE schemes SET params = ?, time_modified = ? WHERE project = ? AND name = ?"
+        cursor.execute(query, (json.dumps(scheme.tags),
+                               datetime.now(), 
+                               self.project_name, 
+                               scheme.name))
+        conn.commit()
+        conn.close()
+        return {"success":"scheme updated"}
+        
+    def delete(self, scheme:SchemeModel):
+        """
+        Delete a scheme
+        """
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "DELETE FROM schemes WHERE project = ? AND name = ?" 
+        cursor.execute(query, (self.project_name,scheme.name))
+        conn.commit()
+        conn.close()
+
+        self.content.drop(columns=scheme.name)
+        self.content.to_parquet(self.path)
+
+        return {"success":"scheme deleted"}
+    
+    def exists(self, name):
+        """
+        Test if scheme exist
+        TODO : harmoniser avec le fichier
+        """
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "SELECT * FROM schemes WHERE project = ? AND name = ?"
+        cursor.execute(query, (self.project_name,name))
+        result = cursor.fetchone()
+        conn.close()
+        if result is None:
+            return False
         else:
-            raise IndexError
+            return True
+
+
+
+    def available(self):
+        """
+        Available schemes
+        """
+        conn = sqlite3.connect(self.db)
+        cursor = conn.cursor()
+        query = "SELECT name, params FROM schemes WHERE project = ?"
+        cursor.execute(query, (self.project_name,))
+        results = cursor.fetchall()
+        conn.commit()
+        conn.close()
+        return {i[0]:json.loads(i[1]) for i in results}
 
     def load(self,json):
         """
@@ -522,10 +664,8 @@ class Schemes():
         self.available = json["available"]    
         self.col = self.col_name()    
     
-    def dump(self):
-        return {
-                "project_name":self.project_name,
-                "name":self.name,
-                "labels":self.labels,
-                "available":self.available
-                }
+    def get(self):
+        return SchemesModel(project_name=self.project_name,
+                            current = self.name,
+                            availables=self.available()
+                            )
